Allow re-registering of specs in Junit for test-retry-plugin
